chore: Move application data to appropriate folder
