+++ apps/civicsignalblog/src/payload/access/applications/main.js
@@ -1,4 +1,4 @@
-import { MAIN } from "../../../lib/data/json/applications";
+import { MAIN } from "../../data/common/applications";
 import canAccessApplication from "../canAccessApplication";
 
 const canRead = ({ req: { user } }) => {
