+++ apps/civicsignalblog/src/payload/access/applications/research.js
@@ -1,4 +1,4 @@
-import { RESEARCH } from "../../../lib/data/json/applications";
+import { RESEARCH } from "../../data/common/applications";
 import canAccessApplication from "../canAccessApplication";
 
 const canRead = ({ req: { user } }) => {
