+++ apps/civicsignalblog/src/payload/data/common/applications.js
